refactor: reorganize and update imports; remove unused nodeSchemas and db files
