--- apps/app/src/db/db.ts
@@ -1,1 +1,0 @@
-export { db } from "@repo/db";
