--- apps/app/src/db/index.ts
@@ -1,1 +1,0 @@
-export * from "./db";
