--- apps/app/src/pages/CanvasPage/nodeSchemas.ts
@@ -1,13 +1,0 @@
-export {
-  ConnectionRuleSchema,
-  OperationNodeDataSchema,
-  PipelineNodeDataSchema,
-  getAllowedConnections,
-  makeDefaultNodeData,
-  makeOperationNodeData,
-  nodeTypeMeta,
-  getNodeMeta,
-} from "./schemas";
-
-export type { NodeType, BuiltinNodeType, OperationNodeData, PipelineNodeData } from "./schemas";
-export type { NodeRunStatus } from "@repo/pipeline-engine/schemas";
