+++ apps/app/src/pages/CanvasPage/schemas/index.ts
@@ -1,8 +1,2 @@
-export { ConnectionRuleSchema } from "@repo/pipeline-engine/schemas";
-export type { NodeType, BuiltinNodeType } from "@repo/pipeline-engine/schemas";
-export { OperationNodeDataSchema, type OperationNodeData } from "./OperationNodeDataSchema";
-export { PipelineNodeDataSchema, type PipelineNodeData } from "./PipelineNodeDataSchema";
-export { getAllowedConnections } from "./getAllowedConnections";
-export { makeDefaultNodeData } from "./makeDefaultNodeData";
-export { makeOperationNodeData } from "./makeOperationNodeData";
-export { nodeTypeMeta, getNodeMeta } from "./nodeTypeMeta";
+export * from "./OperationNodeDataSchema";
+export * from "./PipelineNodeDataSchema";
