+++ apps/app/src/pages/CanvasPage/utils/getAllowedConnections.ts
