+++ apps/app/src/pages/CanvasPage/utils/nodeTypeMeta.ts
