+++ apps/app/src/pages/OperationDetailPage/OperationDetailPageContent/ExecutorCard.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 import { Terminal, Wand2 } from "lucide-react";
 import { SectionHeader } from "../SectionHeader";
-import type { ExecutorConfig } from "../types";
+import type { ExecutorConfig } from "@repo/schemas";
 
 const EXECUTOR_ICON: Record<string, React.ElementType> = {
   agent: Wand2,
