+++ apps/app/src/pages/OperationDetailPage/types.ts
@@ -1,10 +1,3 @@
-export type {
-  PortKind,
-  InputPort,
-  OutputPort,
-  ExecutorConfig,
-  OperationConfig,
-} from "@repo/schemas";
 import type { PortKind } from "@repo/schemas";
 
 export const KIND_LABEL: Record<PortKind, string> = {
