+++ packages/logger/src/index.ts
@@ -1,14 +1,1 @@
-import pino from "pino";
-
-const isDev = process.env.NODE_ENV !== "production";
-
-export const logger = pino({
-  name: "ordine",
-  level: process.env.LOG_LEVEL ?? "info",
-  ...(isDev && {
-    transport: {
-      target: "pino-pretty",
-      options: { colorize: true },
-    },
-  }),
-});
+export * from './pino';
