+++ packages/pipeline-engine/src/schemas/index.ts
@@ -1,7 +1,7 @@
 export * from "./BuiltinNodeTypeSchema";
 export * from "./NodeTypeSchema";
 export * from "./MetaNodeTypeSchema";
-export * from "./resolveMetaType";
+export * from "../resolveMetaType";
 export * from "./NodeRunStatusSchema";
 export * from "./DisclosureModeSchema";
 export * from "./OutputModeSchema";
